fix: update website scale based on page data
